requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2==2.9.9
drf-yasg==1.21.7
django-filter==23.3
django-cors-headers==4.3.0
gunicorn==21.2.0
python-dotenv==1.0.0